separate cast function from mesh bvh node
